+++ apps/web/src/pages/nfts/collections/index.tsx
@@ -1,13 +1,7 @@
 import Collections from 'views/Nft/market/Collections'
-import SubMenu from 'views/Nft/market/components/SubMenu'
 
 const CollectionsPage = () => {
-  return (
-    <>
-      <SubMenu />
-      <Collections />
-    </>
-  )
+  return <Collections />
 }
 
 export default CollectionsPage