+++ apps/web/src/utils/formatTime.ts
@@ -1,5 +1,9 @@
 import dayjs from 'dayjs'
 
 export function formatTime(time: number | Date) {
-  return dayjs(time).format('MMM d, YYYY HH:mm')
+  return dayjs(time).format('MMM D, YYYY HH:mm')
+}
+
+export function formatUnixTime(time: number) {
+  return formatTime(dayjs.unix(time).toDate())
 }